+++ src/com/intellij/idea/plugin/hybris/impex/psi/impl/ImpexAttributeValueMixin.kt
@@ -1,6 +1,6 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
- * Copyright (C) 2019-2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * Copyright (C) 2019-2024 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -26,6 +26,7 @@ import com.intellij.idea.plugin.hybris.impex.psi.ImpexAnyAttributeValue
 import com.intellij.idea.plugin.hybris.impex.psi.ImpexAttribute
 import com.intellij.idea.plugin.hybris.impex.psi.references.ImpexJavaClassBaseReference
 import com.intellij.idea.plugin.hybris.psi.reference.LanguageReference
+import com.intellij.idea.plugin.hybris.system.type.psi.reference.SpringReference
 import com.intellij.lang.ASTNode
 import com.intellij.psi.PsiReference
 import org.jetbrains.kotlin.psi.psiUtil.getParentOfType
@@ -34,6 +35,7 @@ import javax.lang.model.SourceVersion
 
 abstract class ImpexAttributeValueMixin(astNode: ASTNode) : ASTWrapperPsiElement(astNode), ImpexAnyAttributeValue {
 
+    private val disableInterceptorBeansExclusionRegex = "[,'\"]".toRegex()
     private var myReference: PsiReference? = null
 
     override fun getReferences(): Array<PsiReference> = myReference
@@ -60,6 +62,13 @@ abstract class ImpexAttributeValueMixin(astNode: ASTNode) : ASTWrapperPsiElement
             TypeModifier.PROCESSOR -> if (SourceVersion.isName(text)) {
                 ImpexJavaClassBaseReference(this)
             } else null
+
+            // TODO: multi values and wrapped strings are not yet supported.
+            // see https://help.sap.com/docs/SAP_COMMERCE_CLOUD_PUBLIC_CLOUD/aa417173fe4a4ba5a473c93eb730a417/9ce1b60e12714a7dba6ea7e66b4f7acd.html?locale=en-US#disable-interceptors-via-impex
+            TypeModifier.DISABLE_INTERCEPTOR_BEANS -> {
+                if (node.text.contains(disableInterceptorBeansExclusionRegex)) return null
+                else SpringReference(this, node.text)
+            }
 
             else -> null
         }